feat(modal): Allow not to use ajax
This patch offers two things.
1. Allow not to use ajax for forms. Useful for the delete confirmation
   popup
2. If a page does not have a base tag, find the content url from
   $(body).data('base-url')
   The data attribute has been added in plone.app.layout 2.5.1.dev0

Background on the delete confirmation:
We wanted to just redirect on success, but this looses features with ajax.
There is no way to get the statusmessage. The statusmessage is set on the
302 reply, then jquery directly follows the redirect and receives the
information that it shall delete the cookie. Now the success handler has
no trace of the status message. After thinking of patching statusmessage
to maybe add the status message just once as a response header field, I
realized I could just not use ajax on form submit since it offers no
advantage for this use case.